CoinValueBackend/backend_requirements.txt
flask
requests
gunicorn
python-dotenv
flask-cors